internal/repository/postgres_test.go
package repository

import (
	"context"
	"testing"
	"time"

	"github.com/tm-acme-shop/acme-shop-shared-go/models"
)

func TestPostgresOrderRepository_Create(t *testing.T) {
	// TODO(TEAM-PLATFORM): Add integration tests with test database
	t.Skip("Integration test - requires database")

	ctx := context.Background()

	req := &models.CreateOrderRequest{
		UserID: "user_123",
		Items: []models.OrderItem{
			{
				ID:          "item_1",
				ProductID:   "prod_abc",
				ProductName: "Test Product",
				Quantity:    2,
				UnitPrice:   models.Money{Amount: 1000, Currency: "USD"},
				Total:       models.Money{Amount: 2000, Currency: "USD"},
			},
		},
		ShippingAddress: models.Address{
			Line1:      "123 Test St",
			City:       "Test City",
			State:      "TS",
			PostalCode: "12345",
			Country:    "US",
		},
		BillingAddress: models.Address{
			Line1:      "123 Test St",
			City:       "Test City",
			State:      "TS",
			PostalCode: "12345",
			Country:    "US",
		},
	}

	_ = ctx
	_ = req
}

func TestPostgresOrderRepository_GetByID(t *testing.T) {
	// TODO(TEAM-PLATFORM): Add integration tests
	t.Skip("Integration test - requires database")
}

func TestPostgresOrderRepository_UpdateStatus(t *testing.T) {
	// TODO(TEAM-PLATFORM): Add integration tests
	t.Skip("Integration test - requires database")
}

func TestPostgresOrderRepository_List(t *testing.T) {
	// TODO(TEAM-PLATFORM): Add integration tests
	t.Skip("Integration test - requires database")
}

func TestGenerateOrderID(t *testing.T) {
	id := generateOrderID()

	if id == "" {
		t.Error("Expected non-empty order ID")
	}

	if len(id) < 10 {
		t.Errorf("Expected order ID length >= 10, got %d", len(id))
	}

	if id[:4] != "ord_" {
		t.Errorf("Expected order ID to start with 'ord_', got %s", id[:4])
	}
}

func TestLegacyOrder_Conversion(t *testing.T) {
	// TODO(TEAM-API): Remove after migration complete
	legacy := ConvertToLegacyOrder(
		"ord_123",
		"user_456",
		"pending",
		99.99,
		"USD",
	)

	if legacy.Status != "pending" {
		t.Errorf("Expected status 'pending', got %s", legacy.Status)
	}

	if legacy.TotalPrice != 99.99 {
		t.Errorf("Expected total 99.99, got %f", legacy.TotalPrice)
	}

	if legacy.Currency != "USD" {
		t.Errorf("Expected currency 'USD', got %s", legacy.Currency)
	}
}

func TestOrderModel_CanCancel(t *testing.T) {
	tests := []struct {
		name     string
		status   models.OrderStatus
		expected bool
	}{
		{"Pending can cancel", models.OrderStatusPending, true},
		{"Confirmed can cancel", models.OrderStatusConfirmed, true},
		{"Processing cannot cancel", models.OrderStatusProcessing, false},
		{"Shipped cannot cancel", models.OrderStatusShipped, false},
		{"Delivered cannot cancel", models.OrderStatusDelivered, false},
		{"Cancelled cannot cancel", models.OrderStatusCancelled, false},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			order := &models.Order{Status: tt.status}
			if order.CanCancel() != tt.expected {
				t.Errorf("CanCancel() = %v, want %v", order.CanCancel(), tt.expected)
			}
		})
	}
}

func TestOrderModel_CanRefund(t *testing.T) {
	tests := []struct {
		name      string
		status    models.OrderStatus
		paymentID string
		expected  bool
	}{
		{"Delivered with payment can refund", models.OrderStatusDelivered, "pay_123", true},
		{"Delivered without payment cannot refund", models.OrderStatusDelivered, "", false},
		{"Pending cannot refund", models.OrderStatusPending, "pay_123", false},
		{"Shipped cannot refund", models.OrderStatusShipped, "pay_123", false},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			order := &models.Order{
				Status:    tt.status,
				PaymentID: tt.paymentID,
			}
			if order.CanRefund() != tt.expected {
				t.Errorf("CanRefund() = %v, want %v", order.CanRefund(), tt.expected)
			}
		})
	}
}

func TestOrderModel_CalculateTotal(t *testing.T) {
	order := &models.Order{
		Items: []models.OrderItem{
			{Total: models.Money{Amount: 1000, Currency: "USD"}},
			{Total: models.Money{Amount: 2000, Currency: "USD"}},
			{Total: models.Money{Amount: 500, Currency: "USD"}},
		},
		Tax:          models.Money{Amount: 350, Currency: "USD"},
		ShippingCost: models.Money{Amount: 500, Currency: "USD"},
	}

	order.CalculateTotal()

	if order.Subtotal.Amount != 3500 {
		t.Errorf("Expected subtotal 3500, got %d", order.Subtotal.Amount)
	}

	if order.Total.Amount != 4350 {
		t.Errorf("Expected total 4350, got %d", order.Total.Amount)
	}
}

func BenchmarkGenerateOrderID(b *testing.B) {
	for i := 0; i < b.N; i++ {
		generateOrderID()
	}
}

func BenchmarkOrderCalculateTotal(b *testing.B) {
	order := &models.Order{
		Items: []models.OrderItem{
			{Total: models.Money{Amount: 1000, Currency: "USD"}},
			{Total: models.Money{Amount: 2000, Currency: "USD"}},
			{Total: models.Money{Amount: 500, Currency: "USD"}},
		},
		Tax:          models.Money{Amount: 350, Currency: "USD"},
		ShippingCost: models.Money{Amount: 500, Currency: "USD"},
	}

	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		order.CalculateTotal()
	}
}

// Mock helpers for testing
type MockOrderRepository struct {
	orders map[string]*models.Order
}

func NewMockOrderRepository() *MockOrderRepository {
	return &MockOrderRepository{
		orders: make(map[string]*models.Order),
	}
}

func (m *MockOrderRepository) GetByID(ctx context.Context, id string) (*models.Order, error) {
	if order, ok := m.orders[id]; ok {
		return order, nil
	}
	return nil, nil
}

func (m *MockOrderRepository) Create(ctx context.Context, req *models.CreateOrderRequest) (*models.Order, error) {
	order := &models.Order{
		ID:        generateOrderID(),
		UserID:    req.UserID,
		Status:    models.OrderStatusPending,
		Items:     req.Items,
		CreatedAt: time.Now(),
		UpdatedAt: time.Now(),
	}
	m.orders[order.ID] = order
	return order, nil
}
